+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Support for octo-merges (more than 2 parents)
+
 
 ## [0.7.3] - 2026-05-01